refactor: put validation logic in separate files (EC-81)
