+++ ecommerce-app/src/components/LoginForm/LoginForm.tsx
@@ -6,15 +6,7 @@ import Button from "@mui/material/Button";
 import FormControlLabel from "@mui/material/FormControlLabel";
 import Checkbox from "@mui/material/Checkbox";
 import { Formik, Form } from "formik";
-import { object, string } from "yup";
-import {
-  AT_SIGN_DOMAIN_REGEX,
-  UPPERCASE_LETTER_REGEX,
-  LOWERCASE_LETTER_REGEX,
-  DIGIT_REGEX,
-  NO_SPACE_REGEX,
-} from "../../constants/constants";
-import { IFormInitialValues } from "../../types";
+import { initialLoginValues } from "../../constants/constants";
 import { fetchUserLogin } from "../../store/actions/userLoginActions";
 import { UserAuthOptions } from "@commercetools/sdk-client-v2";
 import { useAppDispatch, useAppSelector } from "../../hooks/redux";
@@ -22,51 +14,11 @@ import { convertToUserAuthOptions } from "../../utils/utils";
 import { userLoginClearErrorMessage } from "../../store/slices/userLoginSlice";
 import { Alert, AlertTitle } from "@mui/material";
 import { NavigateFunction, useNavigate } from "react-router-dom";
+import { setLoginSchema } from "../../utils/validation-schemas";
 
-// const existingUser: UserAuthOptions = {
-//   username: "johndoe@example.com",
-//   password: "Secret123",
-// };
 export function LoginForm() {
   const [showPassword, setShowPassword] = React.useState(false);
-  const initialValues: IFormInitialValues = {
-    email: "",
-    password: "",
-    passwordCheck: [],
-  };
-
-  const LoginSchema = object().shape({
-    email: string()
-      .required("Email is required")
-      .trim("Email must not contain leading or trailing whitespace")
-      .strict(true)
-      .matches(NO_SPACE_REGEX, {
-        message: "Email must not contain middle whitespace",
-      })
-      .matches(AT_SIGN_DOMAIN_REGEX, {
-        message: "Email must contain an '@' sign followed by domain in latin",
-      })
-      .email("Email must be properly formatted e.g., user@example.com"),
-
-    password: string()
-      .required("Password is required")
-      .min(8, "Password is too short - should be 8 chars minimum")
-      .trim("Password must not contain leading or trailing whitespace")
-      .strict(true)
-      .matches(UPPERCASE_LETTER_REGEX, {
-        message: "Password must contain at least one latin uppercase letter",
-      })
-      .matches(LOWERCASE_LETTER_REGEX, {
-        message: "Password must contain at least one latin lowercase letter",
-      })
-      .matches(DIGIT_REGEX, {
-        message: "Password must contain at least one digit ",
-      })
-      .matches(NO_SPACE_REGEX, {
-        message: "Password must not contain middle whitespace",
-      }),
-  });
-
+  const LoginSchema = setLoginSchema();
   const { errorMessage, loading, isLogged } = useAppSelector(
     (state) => state.userLogin,
   );
@@ -88,7 +40,7 @@ export function LoginForm() {
 
   return (
     <Formik
-      initialValues={initialValues}
+      initialValues={initialLoginValues}
       validationSchema={LoginSchema}
       onSubmit={(values) => {
         const existingUser: UserAuthOptions = convertToUserAuthOptions(values);
@@ -157,10 +109,7 @@ export function LoginForm() {
                 <FormControlLabel
                   control={
                     <Checkbox
-                      onChange={
-                        () => setShowPassword(!showPassword)
-                        // handleChange
-                      }
+                      onChange={() => setShowPassword(!showPassword)}
                       name="check"
                       disabled={isLogged}
                     />
