+++ ecommerce-app/src/components/SignupForm/SignupForm.tsx
@@ -3,20 +3,11 @@ import TextField from "@mui/material/TextField";
 import FormControl from "@mui/material/FormControl";
 import Button from "@mui/material/Button";
 import { Formik, Form } from "formik";
-import { object, string, lazy } from "yup";
 import {
-  AT_SIGN_DOMAIN_REGEX,
-  UPPERCASE_LETTER_REGEX,
-  LOWERCASE_LETTER_REGEX,
-  DIGIT_REGEX,
-  NO_SPACE_REGEX,
-  NO_SPECIAL_CHARS_REGEX,
-  NO_DIGIT_REGEX,
-  USER_AGE_ALLOWED,
   MAX_HUMAN_AGE,
   countryOptions,
+  initialSignUpValues,
 } from "../../constants/constants";
-import { ISignupInitialValues } from "../../types";
 import { convertToCustomerDraft, subtractYears } from "../../utils/utils";
 import { DatePicker } from "@mui/x-date-pickers";
 import dayjs from "dayjs";
@@ -25,77 +16,7 @@ import { useState } from "react";
 import { CustomerDraft } from "@commercetools/platform-sdk";
 import { useAppDispatch, useAppSelector } from "../../hooks/redux";
 import { fetchUserSignup } from "../../store/actions/userSignupActions";
-
-const postalCodes = require("postal-codes-js");
-
-const initialValues: ISignupInitialValues = {
-  email: "",
-  password: "",
-  firstName: "",
-  lastName: "",
-  dateOfBirth: "",
-  streetNameShipping: "",
-  streetNameBilling: "",
-  cityShipping: "",
-  cityBilling: "",
-  countryShipping: "",
-  countryBilling: "",
-  postalCodeShipping: "",
-  postalCodeBilling: "",
-  passwordCheck: [],
-  commonAddressCheck: [],
-  defaultShippingCheck: [],
-  defaultBillingCheck: [],
-};
-
-const validateName = (field: string) => {
-  return string()
-    .required(`${field} is required`)
-    .trim(`${field} must not contain leading or trailing whitespace`)
-    .strict(true)
-    .matches(NO_SPECIAL_CHARS_REGEX, {
-      message: `${field} must not contain special charachers`,
-    })
-    .matches(NO_DIGIT_REGEX, {
-      message: `${field} must not contain numbers`,
-    });
-};
-
-const validateStreetName = () => {
-  return string()
-    .required("Street is required")
-    .trim("Street must not contain leading or trailing whitespace")
-    .strict(true);
-};
-
-const validateCity = () => {
-  return string()
-    .required("City is required")
-    .trim("City must not contain leading or trailing whitespace")
-    .strict(true)
-    .matches(NO_SPECIAL_CHARS_REGEX, {
-      message: "City must not contain special characters",
-    })
-    .matches(NO_DIGIT_REGEX, {
-      message: "City must not contain numbers",
-    });
-};
-
-const validatePostalCode = (countryCode: string, postalCodeFormat: string) => {
-  return lazy(() =>
-    string()
-      .required("Postal code is required")
-      .test(
-        "is-correct-postal-code",
-        () =>
-          `Postal code must follow the country ${countryCode} format e.g. ${postalCodeFormat}`,
-        (value) => {
-          if (!value || !countryCode) return false;
-          return typeof postalCodes.validate(countryCode, value) === "boolean";
-        },
-      ),
-  );
-};
+import { setSignupSchema } from "../../utils/validation-schemas";
 
 export function SignupForm() {
   const [countryCodeShipping, setCountryCodeShipping] = useState("");
@@ -104,88 +25,15 @@ export function SignupForm() {
   const [postalCodeFormatBilling, setPostalCodeFormatBilling] = useState("");
   const [isCommonAddressChecked, setIsCommonAddressChecked] = useState(false);
 
-  const SignupSchema = object().shape({
-    email: string()
-      .required("Email is required")
-      .trim("Email must not contain leading or trailing whitespace")
-      .strict(true)
-      .matches(NO_SPACE_REGEX, {
-        message: "Email must not contain middle whitespace",
-      })
-      .matches(AT_SIGN_DOMAIN_REGEX, {
-        message: "Email must contain an '@' sign followed by domain in latin",
-      })
-      .email("Email must be properly formatted e.g., user@example.com"),
+  const SchemaOptions = {
+    countryCodeShipping,
+    postalCodeFormatShipping,
+    countryCodeBilling,
+    postalCodeFormatBilling,
+    isCommonAddressChecked,
+  };
 
-    password: string()
-      .required("Password is required")
-      .min(8, "Password is too short - should be 8 chars minimum")
-      .trim("Password must not contain leading or trailing whitespace")
-      .strict(true)
-      .matches(UPPERCASE_LETTER_REGEX, {
-        message: "Password must contain at least one latin uppercase letter",
-      })
-      .matches(LOWERCASE_LETTER_REGEX, {
-        message: "Password must contain at least one latin lowercase letter",
-      })
-      .matches(DIGIT_REGEX, {
-        message: "Password must contain at least one digit ",
-      })
-      .matches(NO_SPACE_REGEX, {
-        message: "Password must not contain middle whitespace",
-      }),
-
-    firstName: validateName("First name"),
-    lastName: validateName("Last name"),
-
-    dateOfBirth: lazy(() =>
-      string()
-        .required("Birthday is required")
-        .test(
-          "is-allowed-age",
-          () =>
-            `You must be ${USER_AGE_ALLOWED} years old or above to register`,
-          (value) => {
-            if (!value) return false;
-            return (
-              new Date(value) < subtractYears(new Date(), USER_AGE_ALLOWED)
-            );
-          },
-        )
-        .test(
-          "is-valid-age",
-          () => `Your age must be within the possible human lifespan`,
-          (value) => {
-            if (!value) return false;
-            return new Date(value) > subtractYears(new Date(), MAX_HUMAN_AGE);
-          },
-        ),
-    ),
-
-    streetNameShipping: validateStreetName(),
-    streetNameBilling: isCommonAddressChecked
-      ? string().notRequired()
-      : validateStreetName(),
-
-    cityShipping: validateCity(),
-    cityBilling: isCommonAddressChecked
-      ? string().notRequired()
-      : validateCity(),
-
-    countryShipping: string().required("Country is required"),
-    countryBilling: isCommonAddressChecked
-      ? string().notRequired()
-      : string().required("Country is required"),
-
-    postalCodeShipping: validatePostalCode(
-      countryCodeShipping,
-      postalCodeFormatShipping,
-    ),
-    postalCodeBilling: isCommonAddressChecked
-      ? string().notRequired()
-      : validatePostalCode(countryCodeBilling, postalCodeFormatBilling),
-  });
-
+  const SignupSchema = setSignupSchema(SchemaOptions);
   const dispatch = useAppDispatch();
   const { loading, isSignedUp, errorMessage } = useAppSelector(
     (state) => state.userSignup,
@@ -193,7 +41,7 @@ export function SignupForm() {
 
   return (
     <Formik
-      initialValues={initialValues}
+      initialValues={initialSignUpValues}
       validationSchema={SignupSchema}
       onSubmit={(values) => {
         const newUser: CustomerDraft = convertToCustomerDraft(values);
