+++ ecommerce-app/src/constants/constants.ts
@@ -1,3 +1,5 @@
+import { IFormInitialValues, ISignupInitialValues } from "../types";
+
 const AT_SIGN_DOMAIN_REGEX =
   /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/u;
 const UPPERCASE_LETTER_REGEX = /[A-Z]/;
@@ -8,10 +10,37 @@ const NO_SPECIAL_CHARS_REGEX = /^(?!.*[()[\]{}*&^%$#@!"+=:;<>?,./\\|_]).*$/;
 const NO_DIGIT_REGEX = /^(?!.*[0-9]).*$/;
 const USER_AGE_ALLOWED = 14;
 const MAX_HUMAN_AGE = 130;
+
 const countryOptions = [
   { label: "USA", countryCode: "US", postalCodeFormat: "20521-9000" },
   { label: "Germany", countryCode: "DE", postalCodeFormat: "12345" },
 ];
+
+const initialLoginValues: IFormInitialValues = {
+  email: "",
+  password: "",
+  passwordCheck: [],
+};
+
+const initialSignUpValues: ISignupInitialValues = {
+  email: "",
+  password: "",
+  firstName: "",
+  lastName: "",
+  dateOfBirth: "",
+  streetNameShipping: "",
+  streetNameBilling: "",
+  cityShipping: "",
+  cityBilling: "",
+  countryShipping: "",
+  countryBilling: "",
+  postalCodeShipping: "",
+  postalCodeBilling: "",
+  passwordCheck: [],
+  commonAddressCheck: [],
+  defaultShippingCheck: [],
+  defaultBillingCheck: [],
+};
 
 export {
   AT_SIGN_DOMAIN_REGEX,
@@ -24,4 +53,6 @@ export {
   USER_AGE_ALLOWED,
   MAX_HUMAN_AGE,
   countryOptions,
+  initialLoginValues,
+  initialSignUpValues,
 };
