+++ ecommerce-app/src/pages/CartPage.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import WorkInProgress from "../components/WorkInProrgess";
 export function CartPage() {
   return <WorkInProgress />;
