+++ ecommerce-app/src/pages/MainPage.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 export function MainPage() {
   return <div>Main</div>;
 }
