+++ ecommerce-app/src/pages/ShopPage.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import WorkInProgress from "../components/WorkInProrgess";
 export function ShopPage() {
   return <WorkInProgress />;
